change loss function

+++ cf-vae/neuvae.py
@@ -73,7 +73,7 @@ class neuVAE:
         self.x_ = placeholder((None, self.input_dim))  # we need these global nodes
         self.x_u_ = placeholder((None, self.user_dim))  # we need these global nodes
         if train:
-            self.rating_ = placeholder((None))
+            self.rating_ = placeholder((None), dtype=tf.uint8)
 
 
         # inference process
